backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
django-cors-headers==4.3.1
python-decouple==3.8
daphne==4.0.0
redis==5.0.1
psycopg2-binary==2.9.7
Pillow==10.0.1
celery==5.3.4
python-dotenv==1.0.0